requirements.txt
Pillow==7.2.0
webcolors==1.11.1
